chore(deps): Remove unused dependencies
Remove some remaining code that was moved to their own crate as well
