+++ evm-ext/src/evm_native/mod.rs
@@ -1,5 +1,5 @@
 pub use self::{
-    native_evm_context::{NativeEVMContext, ResolverBackedDB},
+    native_evm_context::{HeaderForExecution, NativeEVMContext, ResolverBackedDB},
     native_impl::{append_evm_natives, EVM_CALL_FN_NAME},
     state_changes::{extract_evm_changes, genesis_state_changes},
     type_utils::extract_evm_result,
