+++ evm-ext/src/lib.rs
@@ -1,4 +1,1 @@
 pub mod evm_native;
-pub mod native_evm_context;
-pub mod state_changes;
-pub mod type_utils;
