+++ genesis-image/Cargo.toml
@@ -13,51 +13,12 @@ moved-state = { path = "../state" }
 moved-genesis = { path = "../genesis" }
 
 [build-dependencies]
-alloy.workspace = true
-alloy-rlp.workspace = true
-alloy-trie.workspace = true
-aptos-crypto.workspace = true
-aptos-framework.workspace = true
-aptos-gas-algebra.workspace = true
-aptos-gas-meter.workspace = true
-aptos-gas-schedule.workspace = true
-aptos-table-natives.workspace = true
-aptos-types.workspace = true
-aptos-vm.workspace = true
-aptos-vm-types.workspace = true
-aptos-native-interface.workspace = true
-aptos-storage-interface.workspace = true
 bcs.workspace = true
-better_any.workspace = true
-bytes.workspace = true
-eth_trie.workspace = true
-hex.workspace = true
 move-binary-format.workspace = true
 move-core-types.workspace = true
 move-table-extension.workspace = true
-move-vm-runtime.workspace = true
-move-vm-test-utils.workspace = true
-move-vm-types.workspace = true
-once_cell.workspace = true
-op-alloy.workspace = true
-regex.workspace = true
-revm.workspace = true
-serde.workspace = true
-serde_json.workspace = true
-sha2.workspace = true
-sha3.workspace = true
-smallvec.workspace = true
-sui-framework.workspace = true
-sui-types.workspace = true
-thiserror.workspace = true
-tokio.workspace = true
-moved-primitives = { path = "../primitives" }
-moved-evm-ext = { path = "../evm-ext" }
 moved-state = { path = "../state" }
 moved-genesis = { path = "../genesis" }
-anyhow.workspace = true
-move-compiler.workspace = true
-test-case.workspace = true
 
 [features]
 default = []
