+++ moved/src/types/session_id.rs
@@ -4,6 +4,7 @@ use {
     alloy::primitives::U256,
     aptos_types::transaction::EntryFunction,
     aptos_vm::move_vm_ext::UserTransactionContext,
+    moved_evm_ext::evm_native::HeaderForExecution,
     moved_genesis::config::{GenesisConfig, CHAIN_ID},
     moved_primitives::{ToMoveAddress, B256},
 };
@@ -19,7 +20,7 @@ pub struct SessionId {
     pub script_hash: Option<[u8; 32]>,
     pub chain_id: u8,
     pub user_txn_context: Option<UserTransactionContext>,
-    pub block_header: moved_evm_ext::native_evm_context::HeaderForExecution,
+    pub block_header: HeaderForExecution,
 }
 
 impl SessionId {
@@ -28,7 +29,7 @@ impl SessionId {
         maybe_entry_fn: Option<&EntryFunction>,
         tx_hash: &B256,
         genesis_config: &GenesisConfig,
-        block_header: moved_evm_ext::native_evm_context::HeaderForExecution,
+        block_header: HeaderForExecution,
         script_hash: Option<B256>,
     ) -> Self {
         let chain_id = u8_chain_id(genesis_config);
@@ -56,7 +57,7 @@ impl SessionId {
         tx: &DepositedTx,
         tx_hash: &B256,
         genesis_config: &GenesisConfig,
-        block_header: moved_evm_ext::native_evm_context::HeaderForExecution,
+        block_header: HeaderForExecution,
     ) -> Self {
         let chain_id = u8_chain_id(genesis_config);
         let sender = tx.from.to_move_address();
